refactor: unify benchmark scripts
