+++ crosslearner/benchmarks/__init__.py
@@ -1,6 +1,5 @@
 """Benchmarking utilities."""
 
-from .run_benchmarks import run
-from .run_benchmark import run as run_benchmark
+from .run_benchmarks import run, run as run_benchmark, load_external_iris
 
-__all__ = ["run", "run_benchmark"]
+__all__ = ["run", "run_benchmark", "load_external_iris"]
