+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
 [project.scripts]
 crosslearner-train = "crosslearner.__main__:main"
 crosslearner-benchmarks = "crosslearner.benchmarks.run_benchmarks:main"
-crosslearner-benchmark = "crosslearner.benchmarks.run_benchmark:main"
+crosslearner-benchmark = "crosslearner.benchmarks.run_benchmarks:main_baselines"
 
 [tool.setuptools.packages.find]
 where = ["."]
